Add makedirs to FileUtil, move where the directory templates are created Fixes #2004
Attempt to catch the makedirs error on loss of SMB share access and remout with an alias wich user/pass saved to keychane and AppleScript to open the and mount the SMB.

To be added as an argument. For now hard coded.

It works!

retry_all_methods does not work for the calss. tobe worked on.